Broke out some partial classes to make reading this stuff a little easier.
